Move convert time string to spice_utilities
